bot/requirements.txt
aiogram==3.17.0
celery[redis]==5.4.0
redis==5.2.1
asgiref==3.8.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
requests==2.32.3